🧩 Konsolidierte alle WebSocket-Server zu einer einheitlichen Version in ws-server.py – alle relevanten Features übernommen, alte Varianten archiviert
