+++ backend/ws-server/ws-server.py
@@ -1,11 +1,24 @@
 #!/usr/bin/env python3
-"""
-Optimiertes WebSocket Audio-Streaming Backend
-Reduziert Audio-Latenz von ~200ms auf ~50ms durch:
-- Non-blocking async operations
-- Real-time audio streaming
-- Memory-optimierte Verarbeitung
-- Concurrent connection handling
+"""Unified WebSocket Audio Streaming Server.
+
+This server consolidates all historic implementations into a single, modern
+code base.  It combines low latency audio streaming, real-time STT processing
+and runtime switching between Piper and Kokoro TTS engines.  Legacy files
+have been moved to ``archive/`` (`ws-server-pre-tts.py`,
+`ws-server-old.py`).
+
+Features integrated from previous versions:
+
+* Environment based configuration
+* STT audio pre-processing (``numpy``/``soundfile``)
+* Intent routing with optional Flowise/n8n calls (``aiohttp``)
+* Optional debug file saving (``aiofiles``)
+* Metrics API and TTS engine switching
+
+TODO:
+    - Modularize TTS engine switching
+    - Implement authentication/token handling
+
 """
 
 import asyncio
@@ -15,9 +28,9 @@ import base64
 import time
 import uuid
 import numpy as np
-import tempfile
 import os
 import logging
+import io
 from datetime import datetime
 from typing import Dict, Optional, List, AsyncGenerator
 from dataclasses import dataclass, field
@@ -25,32 +38,53 @@ from concurrent.futures import ThreadPoolExecutor
 from collections import deque
 import aiohttp
 import aiofiles
+import soundfile as sf
 from faster_whisper import WhisperModel
+
+# Importiere neues TTS-System
+from .tts import TTSManager, TTSEngineType, TTSConfig
+
 from metrics_api import start_metrics_api
 
 # Enhanced configuration
 @dataclass
 class StreamingConfig:
+    """Configuration read from environment variables."""
+
     # Audio settings optimized for low latency
-    chunk_size: int = 1024  # Smaller chunks = lower latency
-    sample_rate: int = 16000
-    channels: int = 1
-    max_chunk_buffer: int = 50  # Maximum chunks in memory
-    
+    chunk_size: int = int(os.getenv("AUDIO_CHUNK_SIZE", 1024))
+    sample_rate: int = int(os.getenv("SAMPLE_RATE", 16000))
+    channels: int = int(os.getenv("AUDIO_CHANNELS", 1))
+    max_chunk_buffer: int = int(os.getenv("MAX_CHUNK_BUFFER", 50))
+
     # Processing settings
-    stt_workers: int = 2  # Parallel STT processing
-    tts_workers: int = 1
-    max_audio_duration: float = 30.0  # seconds
-    
+    stt_workers: int = int(os.getenv("STT_WORKERS", 2))
+    max_audio_duration: float = float(os.getenv("MAX_AUDIO_DURATION", 30.0))
+
     # WebSocket settings
-    max_connections: int = 100
-    ping_interval: float = 20.0
-    ping_timeout: float = 10.0
-    
+    max_connections: int = int(os.getenv("MAX_CONNECTIONS", 100))
+    ping_interval: float = float(os.getenv("PING_INTERVAL", 20.0))
+    ping_timeout: float = float(os.getenv("PING_TIMEOUT", 10.0))
+    ws_port: int = int(os.getenv("WS_PORT", 8123))
+    metrics_port: int = int(os.getenv("METRICS_PORT", 8124))
+
     # Models
-    stt_model: str = "base"
-    stt_device: str = "cpu"
-    stt_precision: str = "int8"
+    stt_model: str = os.getenv("STT_MODEL", "base")
+    stt_device: str = os.getenv("STT_DEVICE", "cpu")
+    stt_precision: str = os.getenv("STT_PRECISION", "int8")
+
+    # TTS Configuration
+    default_tts_engine: str = os.getenv("DEFAULT_TTS_ENGINE", "piper")  # "piper" or "kokoro"
+    enable_engine_switching: bool = os.getenv("ENABLE_TTS_SWITCHING", "true").lower() == "true"
+
+    # External services
+    flowise_host: str = os.getenv("FLOWISE_HOST", "")
+    flowise_flow_id: str = os.getenv("FLOWISE_FLOW_ID", "")
+    flowise_api_key: str = os.getenv("FLOWISE_API_KEY", "")
+    n8n_url: str = os.getenv("N8N_URL", "")
+
+    # Debugging
+    save_debug_audio: bool = os.getenv("SAVE_DEBUG_AUDIO", "false").lower() == "true"
 
 config = StreamingConfig()
 
@@ -106,8 +140,8 @@ class AudioBuffer:
         return len(self.buffer)
 
 class AsyncSTTEngine:
-    """Non-blocking STT engine with worker pool"""
-    
+    """Non-blocking STT engine with worker pool and preprocessing."""
+
     def __init__(self, model_size: str = "base", device: str = "cpu", workers: int = 2):
         self.model_size = model_size
         self.device = device
@@ -133,115 +167,70 @@ class AsyncSTTEngine:
         )
         
     async def transcribe_audio(self, audio_data: bytes) -> str:
-        """Transcribe audio without blocking event loop"""
+        """Transcribe audio without blocking event loop."""
         if not self.model:
             raise RuntimeError("STT model not initialized")
-            
+
         loop = asyncio.get_event_loop()
-        
+
         try:
+            # Preprocess audio before sending to the worker thread
+            processed = self._preprocess_audio(audio_data)
+
             start_time = time.time()
             result = await loop.run_in_executor(
                 self.executor,
                 self._transcribe_sync,
-                audio_data
+                processed
             )
-            
+
             processing_time = time.time() - start_time
             logger.debug(f"STT processing took {processing_time:.2f}s")
-            
+
             return result
-            
+
         except Exception as e:
             logger.error(f"STT transcription failed: {e}")
             return f"[STT Error: {str(e)}]"
-            
-    def _transcribe_sync(self, audio_data: bytes) -> str:
-        """Synchronous transcription in worker thread"""
+
+    def _preprocess_audio(self, audio_bytes: bytes) -> np.ndarray:
+        """Normalize and resample audio bytes for STT."""
+        audio_buf = io.BytesIO(audio_bytes)
+        data, sr = sf.read(audio_buf, dtype="float32")
+
+        # Resample if necessary
+        if sr != config.sample_rate:
+            duration = data.shape[0] / sr
+            target_length = int(duration * config.sample_rate)
+            data = np.interp(
+                np.linspace(0, data.shape[0], target_length),
+                np.arange(data.shape[0]),
+                data,
+            )
+
+        # Simple normalization
+        max_val = np.max(np.abs(data))
+        if max_val > 0:
+            data = data / max_val
+
+        return data.astype(np.float32)
+
+    def _transcribe_sync(self, audio_array: np.ndarray) -> str:
+        """Synchronous transcription in worker thread."""
         try:
-            # Save to temporary file for whisper
-            with tempfile.NamedTemporaryFile(delete=False, suffix=".wav") as tmp_file:
-                tmp_file.write(audio_data)
-                tmp_path = tmp_file.name
-                
-            # Transcribe
-            segments, info = self.model.transcribe(tmp_path)
+            segments, info = self.model.transcribe(audio_array, sample_rate=config.sample_rate)
             text = "".join(segment.text for segment in segments).strip()
-            
-            # Cleanup
-            os.unlink(tmp_path)
-            
             return text or "(no speech detected)"
-            
         except Exception as e:
             logger.error(f"Sync transcription error: {e}")
             return f"[Transcription failed: {str(e)}]"
 
-class AsyncTTSEngine:
-    """Non-blocking TTS engine"""
-    
-    def __init__(self, model_path: str = None, workers: int = 1):
-        self.model_path = model_path or os.path.expanduser("~/.local/share/piper/de-thorsten-low.onnx")
-        self.executor = ThreadPoolExecutor(max_workers=workers, thread_name_prefix="TTS")
-        
-    async def synthesize(self, text: str, voice: str = "de-thorsten") -> Optional[bytes]:
-        """Generate speech audio without blocking"""
-        if not text.strip():
-            return None
-            
-        loop = asyncio.get_event_loop()
-        
-        try:
-            start_time = time.time()
-            audio_data = await loop.run_in_executor(
-                self.executor,
-                self._synthesize_sync,
-                text
-            )
-            
-            processing_time = time.time() - start_time
-            logger.debug(f"TTS processing took {processing_time:.2f}s")
-            
-            return audio_data
-            
-        except Exception as e:
-            logger.error(f"TTS synthesis failed: {e}")
-            return None
-            
-    def _synthesize_sync(self, text: str) -> Optional[bytes]:
-        """Synchronous TTS in worker thread"""
-        try:
-            with tempfile.NamedTemporaryFile(delete=False, suffix=".wav") as output_file:
-                output_path = output_file.name
-                
-            # Run piper TTS
-            import subprocess
-            process = subprocess.run([
-                'piper',
-                '--model', self.model_path,
-                '--output_file', output_path,
-                '--text', text
-            ], capture_output=True, text=True, timeout=10)
-            
-            if process.returncode == 0 and os.path.exists(output_path):
-                with open(output_path, 'rb') as f:
-                    audio_data = f.read()
-                os.unlink(output_path)
-                return audio_data
-            else:
-                logger.error(f"Piper TTS failed: {process.stderr}")
-                return None
-                
-        except Exception as e:
-            logger.error(f"TTS synthesis error: {e}")
-            return None
-
 class AudioStreamManager:
     """Manages real-time audio streams with minimal latency"""
     
-    def __init__(self, stt_engine: AsyncSTTEngine, tts_engine: AsyncTTSEngine):
+    def __init__(self, stt_engine: AsyncSTTEngine, tts_manager: TTSManager):
         self.stt_engine = stt_engine
-        self.tts_engine = tts_engine
+        self.tts_manager = tts_manager
         self.active_streams: Dict[str, Dict] = {}
         self.processing_queue = asyncio.Queue(maxsize=1000)
         self.response_callbacks: Dict[str, callable] = {}
@@ -259,7 +248,9 @@ class AudioStreamManager:
             'start_time': time.time(),
             'last_activity': time.time(),
             'is_active': True,
-            'chunk_count': 0
+            'chunk_count': 0,
+            'tts_engine': None,  # Client-spezifische TTS-Engine
+            'tts_voice': None    # Client-spezifische Stimme
         }
         
         self.response_callbacks[stream_id] = response_callback
@@ -318,7 +309,9 @@ class AudioStreamManager:
                     'stream_id': stream_id,
                     'audio_data': audio_data,
                     'client_id': stream['client_id'],
-                    'chunk_count': stream['chunk_count']
+                    'chunk_count': stream['chunk_count'],
+                    'tts_engine': stream.get('tts_engine'),
+                    'tts_voice': stream.get('tts_voice')
                 })
                 logger.debug(f"Queued stream {stream_id} for processing ({len(audio_data)} bytes)")
                 return True
@@ -329,6 +322,14 @@ class AudioStreamManager:
             logger.warning(f"No audio data in finalized stream {stream_id}")
             return False
             
+    async def set_stream_tts_config(self, stream_id: str, engine: Optional[str] = None, voice: Optional[str] = None):
+        """Setze TTS-Konfiguration für Stream"""
+        if stream_id in self.active_streams:
+            if engine:
+                self.active_streams[stream_id]['tts_engine'] = engine
+            if voice:
+                self.active_streams[stream_id]['tts_voice'] = voice
+                
     async def _process_audio_queue(self):
         """Background processor for audio streams"""
         while True:
@@ -348,18 +349,44 @@ class AudioStreamManager:
         stream_id = item['stream_id']
         audio_data = item['audio_data']
         client_id = item['client_id']
+        tts_engine = item.get('tts_engine')
+        tts_voice = item.get('tts_voice')
         
         try:
             start_time = time.time()
-            
+
             # Transcribe audio
             transcription = await self.stt_engine.transcribe_audio(audio_data)
             
             # Generate response (simplified - would include intent routing)
             response_text = await self._generate_response(transcription, client_id)
             
-            # Generate TTS audio
-            tts_audio = await self.tts_engine.synthesize(response_text)
+            # TTS-Engine bestimmen
+            target_engine = None
+            if tts_engine:
+                if tts_engine.lower() == "piper":
+                    target_engine = TTSEngineType.PIPER
+                elif tts_engine.lower() == "kokoro":
+                    target_engine = TTSEngineType.KOKORO
+                    
+            # Generate TTS audio mit spezifizierter Engine
+            tts_result = await self.tts_manager.synthesize(
+                response_text,
+                engine=target_engine,
+                voice=tts_voice
+            )
+
+            # Optional debug: save audio files asynchronously
+            if config.save_debug_audio:
+                timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+                try:
+                    async with aiofiles.open(f"debug_in_{timestamp}.wav", "wb") as f_in:
+                        await f_in.write(audio_data)
+                    if tts_result.success and tts_result.audio_data:
+                        async with aiofiles.open(f"debug_out_{timestamp}.wav", "wb") as f_out:
+                            await f_out.write(tts_result.audio_data)
+                except Exception as dbg_err:
+                    logger.debug(f"Failed to write debug audio: {dbg_err}")
             
             processing_time = time.time() - start_time
             logger.info(f"Processed stream {stream_id} in {processing_time:.2f}s: '{transcription[:50]}...'")
@@ -371,8 +398,12 @@ class AudioStreamManager:
                     'type': 'audio_response',
                     'transcription': transcription,
                     'response_text': response_text,
-                    'audio_data': tts_audio,
+                    'audio_data': tts_result.audio_data if tts_result.success else None,
                     'processing_time_ms': round(processing_time * 1000),
+                    'tts_engine_used': tts_result.engine_used,
+                    'tts_voice_used': tts_result.voice_used,
+                    'tts_success': tts_result.success,
+                    'tts_error': tts_result.error_message if not tts_result.success else None,
                     'stream_id': stream_id
                 })
                 
@@ -387,31 +418,69 @@ class AudioStreamManager:
                 del self.response_callbacks[stream_id]
                 
     async def _generate_response(self, transcription: str, client_id: str) -> str:
-        """Generate response to transcription (simplified)"""
-        # This would include intent routing, external API calls, etc.
-        # For now, simple echo with timestamp
+        """Generate response to transcription via simple intent routing."""
         if not transcription or transcription.startswith('['):
             return "Entschuldigung, ich konnte Sie nicht verstehen."
-            
-        # Simple responses for common phrases
+
         text = transcription.lower().strip()
-        
+
+        # Local shortcuts
         if any(word in text for word in ['zeit', 'uhrzeit', 'wie spät']):
             return f"Es ist {datetime.now().strftime('%H:%M')} Uhr."
-        elif any(word in text for word in ['hallo', 'hi', 'guten tag']):
+        if any(word in text for word in ['hallo', 'hi', 'guten tag']):
             return "Hallo! Wie kann ich Ihnen helfen?"
-        elif any(word in text for word in ['danke', 'vielen dank']):
+        if any(word in text for word in ['danke', 'vielen dank']):
             return "Gern geschehen!"
-        else:
-            return f"Sie sagten: {transcription}"
+        if any(word in text for word in ['stimme', 'voice', 'engine']):
+            current_engine = self.tts_manager.get_current_engine()
+            return f"Ich verwende aktuell {current_engine.value if current_engine else 'keine'} TTS-Engine."
+
+        # External routing examples
+        if config.n8n_url and any(word in text for word in ['garage', 'status', 'wetter']):
+            return await self._trigger_n8n(text)
+        if config.flowise_host and any(word in text for word in ['frage', 'wissen', 'hilfe']):
+            return await self._ask_flowise(text)
+
+        return f"Sie sagten: {transcription}"
+
+    async def _ask_flowise(self, query: str) -> str:
+        """Send query to Flowise REST API."""
+        url = f"{config.flowise_host}/api/v1/prediction/{config.flowise_flow_id}"
+        headers = {"Content-Type": "application/json"}
+        if config.flowise_api_key:
+            headers["Authorization"] = f"Bearer {config.flowise_api_key}"
+        try:
+            async with aiohttp.ClientSession() as session:
+                async with session.post(url, headers=headers, json={"question": query}) as resp:
+                    if resp.status == 200:
+                        data = await resp.json()
+                        return data.get("text", "(keine Antwort von Flowise)")
+                    return f"[Flowise Fehler {resp.status}]"
+        except Exception as e:
+            logger.error(f"Flowise request failed: {e}")
+            return f"[Flowise Fehler: {e}]"
+
+    async def _trigger_n8n(self, query: str) -> str:
+        """Trigger n8n webhook with the given query."""
+        try:
+            async with aiohttp.ClientSession() as session:
+                async with session.post(config.n8n_url, json={"query": query}) as resp:
+                    if resp.status == 200:
+                        data = await resp.json()
+                        return data.get("reply", "OK, erledigt")
+                    return f"[n8n Fehler {resp.status}]"
+        except Exception as e:
+            logger.error(f"n8n request failed: {e}")
+            return f"[n8n Fehler: {e}]"
 
 class ConnectionManager:
     """Manages WebSocket connections with optimized handling"""
     
-    def __init__(self, stream_manager: AudioStreamManager):
+    def __init__(self, stream_manager: AudioStreamManager, tts_manager: TTSManager):
         self.active_connections: Dict[str, websockets.WebSocketServerProtocol] = {}
         self.connection_info: Dict[str, Dict] = {}
         self.stream_manager = stream_manager
+        self.tts_manager = tts_manager
         
     async def register(self, websocket: websockets.WebSocketServerProtocol) -> str:
         """Register new WebSocket connection"""
@@ -423,7 +492,9 @@ class ConnectionManager:
             'last_activity': time.time(),
             'remote_addr': websocket.remote_address[0] if websocket.remote_address else 'unknown',
             'messages_sent': 0,
-            'messages_received': 0
+            'messages_received': 0,
+            'preferred_tts_engine': config.default_tts_engine,
+            'preferred_tts_voice': None
         }
         
         logger.info(f"Client {client_id} connected from {websocket.remote_address}")
@@ -462,7 +533,7 @@ class ConnectionManager:
             return False
 
 class OptimizedVoiceServer:
-    """Main server with optimized audio streaming"""
+    """Main server with optimized audio streaming and TTS switching"""
     
     def __init__(self):
         self.stt_engine = AsyncSTTEngine(
@@ -470,25 +541,56 @@ class OptimizedVoiceServer:
             device=config.stt_device,
             workers=config.stt_workers
         )
-        self.tts_engine = AsyncTTSEngine(workers=config.tts_workers)
-        self.stream_manager = AudioStreamManager(self.stt_engine, self.tts_engine)
-        self.connection_manager = ConnectionManager(self.stream_manager)
+        
+        # Initialisiere TTS-Manager
+        self.tts_manager = TTSManager()
+        
+        self.stream_manager = AudioStreamManager(self.stt_engine, self.tts_manager)
+        self.connection_manager = ConnectionManager(self.stream_manager, self.tts_manager)
         
         # Performance metrics
         self.stats = {
             'connections': 0,
             'messages_processed': 0,
             'audio_streams': 0,
+            'tts_switches': 0,
             'start_time': time.time()
         }
         
     async def initialize(self):
         """Initialize all components"""
-        logger.info("Initializing optimized voice server...")
+        logger.info("Initializing optimized voice server with TTS switching...")
         
         # Initialize STT engine
         await self.stt_engine.initialize()
         
+        # Initialize TTS Manager
+        # Standard-Konfigurationen für beide Engines
+        piper_config = TTSConfig(
+            engine_type="piper",
+            model_path="",  # Wird automatisch ermittelt
+            voice="de-thorsten-low",
+            speed=1.0,
+            language="de",
+            sample_rate=22050
+        )
+        
+        kokoro_config = TTSConfig(
+            engine_type="kokoro",
+            model_path="",  # Wird automatisch ermittelt
+            voice="af_sarah",
+            speed=1.0,
+            language="en",
+            sample_rate=24000
+        )
+        
+        # Bestimme Standard-Engine
+        default_engine = TTSEngineType.PIPER if config.default_tts_engine.lower() == "piper" else TTSEngineType.KOKORO
+        
+        success = await self.tts_manager.initialize(piper_config, kokoro_config, default_engine)
+        if not success:
+            logger.error("TTS-Manager Initialisierung fehlgeschlagen!")
+            
         logger.info("Voice server initialized successfully")
         
     async def handle_websocket(self, websocket, path):
@@ -496,7 +598,10 @@ class OptimizedVoiceServer:
         client_id = await self.connection_manager.register(websocket)
         
         try:
-            # Send welcome message
+            # Send welcome message with TTS info
+            available_engines = await self.tts_manager.get_available_engines()
+            current_engine = self.tts_manager.get_current_engine()
+            
             await self.connection_manager.send_to_client(client_id, {
                 'type': 'connected',
                 'client_id': client_id,
@@ -504,7 +609,13 @@ class OptimizedVoiceServer:
                 'config': {
                     'chunk_size': config.chunk_size,
                     'sample_rate': config.sample_rate,
-                    'max_duration': config.max_audio_duration
+                    'max_duration': config.max_audio_duration,
+                    'tts_switching_enabled': config.enable_engine_switching
+                },
+                'tts_info': {
+                    'available_engines': available_engines,
+                    'current_engine': current_engine.value if current_engine else None,
+                    'switching_enabled': config.enable_engine_switching
                 }
             })
             
@@ -544,6 +655,14 @@ class OptimizedVoiceServer:
             await self._handle_end_audio_stream(client_id, data)
         elif message_type == 'text':
             await self._handle_text_message(client_id, data)
+        elif message_type == 'switch_tts_engine':
+            await self._handle_switch_tts_engine(client_id, data)
+        elif message_type == 'get_tts_info':
+            await self._handle_get_tts_info(client_id, data)
+        elif message_type == 'set_tts_voice':
+            await self._handle_set_tts_voice(client_id, data)
+        elif message_type == 'test_tts_engines':
+            await self._handle_test_tts_engines(client_id, data)
         elif message_type == 'ping':
             await self._handle_ping(client_id, data)
         else:
@@ -554,15 +673,24 @@ class OptimizedVoiceServer:
             
     async def _handle_start_audio_stream(self, client_id: str, data: Dict):
         """Start new audio stream"""
+        # TTS-Konfiguration aus Client-Einstellungen
+        tts_engine = data.get('tts_engine')
+        tts_voice = data.get('tts_voice')
+        
         # Create response callback
         async def response_callback(response_data):
             await self._send_audio_response(client_id, response_data)
             
         stream_id = await self.stream_manager.start_stream(client_id, response_callback)
         
+        # TTS-Konfiguration für Stream setzen
+        await self.stream_manager.set_stream_tts_config(stream_id, tts_engine, tts_voice)
+        
         await self.connection_manager.send_to_client(client_id, {
             'type': 'audio_stream_started',
             'stream_id': stream_id,
+            'tts_engine': tts_engine,
+            'tts_voice': tts_voice,
             'timestamp': time.time()
         })
         
@@ -613,27 +741,173 @@ class OptimizedVoiceServer:
     async def _handle_text_message(self, client_id: str, data: Dict):
         """Handle text-only message"""
         text = data.get('content', '').strip()
+        tts_engine = data.get('tts_engine')
+        tts_voice = data.get('tts_voice')
+        
         if not text:
             return
             
         # Generate response
         response_text = await self.stream_manager._generate_response(text, client_id)
         
+        # TTS-Engine bestimmen
+        target_engine = None
+        if tts_engine:
+            if tts_engine.lower() == "piper":
+                target_engine = TTSEngineType.PIPER
+            elif tts_engine.lower() == "kokoro":
+                target_engine = TTSEngineType.KOKORO
+                
         # Generate TTS
-        tts_audio = await self.tts_engine.synthesize(response_text)
+        tts_result = await self.tts_manager.synthesize(
+            response_text, 
+            engine=target_engine,
+            voice=tts_voice
+        )
         
         await self._send_text_response(client_id, {
             'input_text': text,
             'response_text': response_text,
-            'audio_data': tts_audio
+            'audio_data': tts_result.audio_data if tts_result.success else None,
+            'tts_engine_used': tts_result.engine_used,
+            'tts_voice_used': tts_result.voice_used,
+            'tts_success': tts_result.success,
+            'tts_error': tts_result.error_message if not tts_result.success else None
+        })
+        
+    async def _handle_switch_tts_engine(self, client_id: str, data: Dict):
+        """Handle TTS engine switch request"""
+        if not config.enable_engine_switching:
+            await self.connection_manager.send_to_client(client_id, {
+                'type': 'tts_switch_error',
+                'message': 'TTS engine switching is disabled'
+            })
+            return
+            
+        engine_name = data.get('engine', '').lower()
+        
+        # Engine-Type bestimmen
+        if engine_name == 'piper':
+            target_engine = TTSEngineType.PIPER
+        elif engine_name == 'kokoro':
+            target_engine = TTSEngineType.KOKORO
+        else:
+            await self.connection_manager.send_to_client(client_id, {
+                'type': 'tts_switch_error',
+                'message': f'Unknown engine: {engine_name}'
+            })
+            return
+            
+        # Engine wechseln
+        success = await self.tts_manager.switch_engine(target_engine)
+        
+        if success:
+            self.stats['tts_switches'] += 1
+            
+            # Update client preferences
+            self.connection_manager.connection_info[client_id]['preferred_tts_engine'] = engine_name
+            
+            await self.connection_manager.send_to_client(client_id, {
+                'type': 'tts_engine_switched',
+                'engine': engine_name,
+                'timestamp': time.time()
+            })
+        else:
+            await self.connection_manager.send_to_client(client_id, {
+                'type': 'tts_switch_error',
+                'message': f'Failed to switch to {engine_name}'
+            })
+            
+    async def _handle_get_tts_info(self, client_id: str, data: Dict):
+        """Handle TTS info request"""
+        available_engines = await self.tts_manager.get_available_engines()
+        available_voices = await self.tts_manager.get_available_voices()
+        current_engine = self.tts_manager.get_current_engine()
+        engine_stats = self.tts_manager.get_engine_stats()
+        
+        await self.connection_manager.send_to_client(client_id, {
+            'type': 'tts_info',
+            'available_engines': available_engines,
+            'available_voices': available_voices,
+            'current_engine': current_engine.value if current_engine else None,
+            'engine_stats': engine_stats,
+            'switching_enabled': config.enable_engine_switching,
+            'timestamp': time.time()
+        })
+        
+    async def _handle_set_tts_voice(self, client_id: str, data: Dict):
+        """Handle TTS voice change request"""
+        voice = data.get('voice')
+        engine = data.get('engine')  # Optional: spezifische Engine
+        
+        if not voice:
+            await self.connection_manager.send_to_client(client_id, {
+                'type': 'tts_voice_error',
+                'message': 'No voice specified'
+            })
+            return
+            
+        # Engine bestimmen
+        target_engine = None
+        if engine:
+            if engine.lower() == 'piper':
+                target_engine = TTSEngineType.PIPER
+            elif engine.lower() == 'kokoro':
+                target_engine = TTSEngineType.KOKORO
+                
+        success = await self.tts_manager.set_voice(voice, target_engine)
+        
+        if success:
+            # Update client preferences
+            self.connection_manager.connection_info[client_id]['preferred_tts_voice'] = voice
+            
+            await self.connection_manager.send_to_client(client_id, {
+                'type': 'tts_voice_changed',
+                'voice': voice,
+                'engine': engine or self.tts_manager.get_current_engine().value,
+                'timestamp': time.time()
+            })
+        else:
+            await self.connection_manager.send_to_client(client_id, {
+                'type': 'tts_voice_error',
+                'message': f'Failed to set voice: {voice}'
+            })
+            
+    async def _handle_test_tts_engines(self, client_id: str, data: Dict):
+        """Handle TTS engine test request"""
+        test_text = data.get('text', 'Test der Sprachsynthese')
+        
+        results = await self.tts_manager.test_all_engines(test_text)
+        
+        # Convert results for JSON serialization
+        serializable_results = {}
+        for engine_name, result in results.items():
+            serializable_results[engine_name] = {
+                'success': result.success,
+                'processing_time_ms': result.processing_time_ms,
+                'voice_used': result.voice_used,
+                'engine_used': result.engine_used,
+                'error_message': result.error_message,
+                'audio_length_ms': result.audio_length_ms,
+                'sample_rate': result.sample_rate
+            }
+            
+        await self.connection_manager.send_to_client(client_id, {
+            'type': 'tts_test_results',
+            'results': serializable_results,
+            'test_text': test_text,
+            'timestamp': time.time()
         })
         
     async def _handle_ping(self, client_id: str, data: Dict):
         """Handle ping message"""
+        current_engine = self.tts_manager.get_current_engine()
+        
         await self.connection_manager.send_to_client(client_id, {
             'type': 'pong',
             'timestamp': time.time(),
-            'client_timestamp': data.get('timestamp')
+            'client_timestamp': data.get('timestamp'),
+            'current_tts_engine': current_engine.value if current_engine else None
         })
         
     async def _send_audio_response(self, client_id: str, response_data: Dict):
@@ -650,6 +924,10 @@ class OptimizedVoiceServer:
             'content': response_data.get('response_text'),
             'audio': f"data:audio/wav;base64,{audio_b64}" if audio_b64 else None,
             'processing_time_ms': response_data.get('processing_time_ms'),
+            'tts_engine_used': response_data.get('tts_engine_used'),
+            'tts_voice_used': response_data.get('tts_voice_used'),
+            'tts_success': response_data.get('tts_success'),
+            'tts_error': response_data.get('tts_error'),
             'timestamp': time.time()
         })
         
@@ -665,6 +943,10 @@ class OptimizedVoiceServer:
             'type': 'response',
             'content': response_data.get('response_text'),
             'audio': f"data:audio/wav;base64,{audio_b64}" if audio_b64 else None,
+            'tts_engine_used': response_data.get('tts_engine_used'),
+            'tts_voice_used': response_data.get('tts_voice_used'),
+            'tts_success': response_data.get('tts_success'),
+            'tts_error': response_data.get('tts_error'),
             'timestamp': time.time()
         })
         
@@ -675,9 +957,11 @@ class OptimizedVoiceServer:
             'total_connections': self.stats['connections'],
             'messages_processed': self.stats['messages_processed'],
             'audio_streams_processed': self.stats['audio_streams'],
+            'tts_engine_switches': self.stats['tts_switches'],
             'uptime_seconds': time.time() - self.stats['start_time'],
             'active_audio_streams': len(self.stream_manager.active_streams),
-            'processing_queue_size': self.stream_manager.processing_queue.qsize()
+            'processing_queue_size': self.stream_manager.processing_queue.qsize(),
+            'tts_engines': self.tts_manager.get_engine_stats()
         }
 
 # Main server instance
@@ -690,32 +974,34 @@ async def main():
     
     # Start HTTP Metrics API
     try:
-        metrics_runner = await start_metrics_api(server, port=8124)
-        logger.info("📊 Metrics API started on port 8124")
+        metrics_runner = await start_metrics_api(server, port=config.metrics_port)
+        logger.info(f"📊 Metrics API started on port {config.metrics_port}")
     except Exception as e:
         logger.warning(f"Failed to start metrics API: {e}")
         metrics_runner = None
     
     # Start WebSocket server
-    logger.info(f"Starting optimized WebSocket server on port 8123")
+    logger.info(f"Starting optimized WebSocket server with TTS switching on port {config.ws_port}")
     
     try:
         async with websockets.serve(
             server.handle_websocket,
             "0.0.0.0",
-            8123,
+            config.ws_port,
             max_size=10_000_000,  # 10MB max message size
             ping_interval=config.ping_interval,
             ping_timeout=config.ping_timeout,
             close_timeout=10
         ):
-            logger.info("🚀 Optimized Voice Server is running!")
-            logger.info("📊 Metrics available at: http://localhost:8124/metrics")
-            logger.info("🏥 Health check at: http://localhost:8124/health")
+            logger.info("🚀 Optimized Voice Server with TTS switching is running!")
+            logger.info(f"📊 Metrics available at: http://localhost:{config.metrics_port}/metrics")
+            logger.info(f"🏥 Health check at: http://localhost:{config.metrics_port}/health")
+            logger.info("🎙️  TTS Engine switching enabled" if config.enable_engine_switching else "🎙️  TTS Engine switching disabled")
             await asyncio.Future()  # Run forever
     finally:
         if metrics_runner:
             await metrics_runner.cleanup()
+        await server.tts_manager.cleanup()
 
 if __name__ == "__main__":
     try:
